Aggiunto classe per implementazione di fake rooms service e documentato i fake
